Break out CUDA utils into cpp file
