+++ pennylane_lightning/core/bindings/BindingsCudaUtils_nb.cpp
@@ -1,0 +1,77 @@
+#include "BindingsUtils_nb.hpp"
+#include "DevTag.hpp"
+#include "DevicePool.hpp"
+#include "cuda_helpers.hpp"
+
+namespace Pennylane::LightningGPU::Util::NanoBindings {
+
+namespace nb = nanobind;
+
+/**
+ * @brief Register bindings for CUDA utils.
+ *
+ * @param m Nanobind module.
+ */
+void registerCudaUtils(nb::module_ &m) {
+    m.def("device_reset", &deviceReset, "Reset all GPU devices and contexts.");
+    m.def("allToAllAccess", []() {
+        for (int i = 0; i < static_cast<int>(getGPUCount()); i++) {
+            cudaDeviceEnablePeerAccess(i, 0);
+        }
+    });
+
+    m.def("is_gpu_supported", &isCuQuantumSupported,
+          nb::arg("device_number") = 0,
+          "Checks if the given GPU device meets the minimum architecture "
+          "support for the PennyLane-Lightning-GPU device.");
+
+    m.def("get_gpu_arch", &getGPUArch, nb::arg("device_number") = 0,
+          "Returns the given GPU major and minor GPU support.");
+
+    auto pyclass_devpool = nb::class_<DevicePool<int>>(m, "DevPool");
+    pyclass_devpool.def(nb::init<>());
+    pyclass_devpool.def("getActiveDevices", &DevicePool<int>::getActiveDevices);
+    pyclass_devpool.def("isActive", &DevicePool<int>::isActive);
+    pyclass_devpool.def("isInactive", &DevicePool<int>::isInactive);
+    pyclass_devpool.def("acquireDevice", &DevicePool<int>::acquireDevice);
+    pyclass_devpool.def("releaseDevice", &DevicePool<int>::releaseDevice);
+    pyclass_devpool.def("syncDevice", &DevicePool<int>::syncDevice);
+    pyclass_devpool.def("refresh", &DevicePool<int>::refresh);
+    pyclass_devpool.def_static("getTotalDevices",
+                               &DevicePool<int>::getTotalDevices);
+    pyclass_devpool.def_static("getDeviceUIDs",
+                               &DevicePool<int>::getDeviceUIDs);
+    pyclass_devpool.def_static("setDeviceID", &DevicePool<int>::setDeviceIdx);
+    pyclass_devpool.def(
+        "__getstate__",
+        []([[maybe_unused]] const DevicePool<int> &self) { // __getstate__
+            return nb::make_tuple();
+        });
+    pyclass_devpool.def("__setstate__", [](DevicePool<int> &self,
+                                           nb::tuple &t) { // __setstate__
+        if (t.size() != 0) {
+            throw std::runtime_error("Invalid state!");
+        }
+
+        new (&self) DevicePool<int>(); // Reconstruct the object
+    });
+
+    auto pyclass_devtag = nb::class_<DevTag<int>>(m, "DevTag");
+    pyclass_devtag.def(nb::init<>());
+    pyclass_devtag.def(nb::init<int>());
+    pyclass_devtag.def(nb::init<const DevTag<int> &>());
+    pyclass_devtag.def("__init__", [](int device_id, void *stream_id) {
+        // The lower level `__init__` needs to be defined directly to
+        // support type casting
+        return DevTag<int>(device_id, static_cast<cudaStream_t>(stream_id));
+    });
+    pyclass_devtag.def("getDeviceID", &DevTag<int>::getDeviceID);
+    pyclass_devtag.def("getStreamID", [](DevTag<int> &dev_tag) {
+        // default stream points to nullptr, so just return void* as
+        // type
+        return static_cast<void *>(dev_tag.getStreamID());
+    });
+    pyclass_devtag.def("refresh", &DevTag<int>::refresh);
+}
+
+} // namespace Pennylane::LightningGPU::Util::NanoBindings
