+++ pennylane_lightning/core/simulators/lightning_gpu/bindings/LGPUBindings_nb.hpp
@@ -33,6 +33,8 @@
 #include "StateVectorCudaManaged.hpp"
 #include "TypeList.hpp"
 
+namespace nb = nanobind;
+
 /// @cond DEV
 namespace {
 using namespace Pennylane::LightningGPU::Measures;
@@ -41,8 +43,6 @@ using namespace Pennylane::LightningGPU::Observables;
 /// @endcond
 
 namespace Pennylane::LightningGPU::NanoBindings {
-
-namespace nb = nanobind;
 
 /**
  * @brief Define StateVector backends for lightning.gpu
@@ -236,7 +236,7 @@ void registerBackendSpecificInfo(nb::module_ &m) {
             return info;
         },
         "Backend-specific information.");
-    registerCudaUtils(m);
+    Pennylane::LightningGPU::Util::NanoBindings::registerCudaUtils(m);
 } // m
 
 /**
